fixed and filled out many of the type declarations for all of the new types removed obsolete restart_ptl.[ch]

+++ adios.c
@@ -2,6 +2,7 @@
 #include <math.h>
 #include <string.h>
 #include <unistd.h>
+#include <stdint.h>
 
 // xml parser
 #include <mxml.h>
@@ -267,7 +268,7 @@ void adios_write_ (long long * fd_p, const char * name, void * var, int err
 
 ///////////////////////////////////////////////////////////////////////////////
 static int common_adios_get_write_buffer (long long fd_p, const char * name
-                                         ,unsigned long long * size
+                                         ,uint64_t * size
                                          ,void ** buffer
                                          )
 {
@@ -318,7 +319,7 @@ static int common_adios_get_write_buffer (long long fd_p, const char * name
 }
 
 int adios_get_write_buffer (long long fd_p, const char * name
-                           ,unsigned long long * size
+                           ,uint64_t * size
                            ,void ** buffer
                            )
 {
@@ -326,7 +327,7 @@ int adios_get_write_buffer (long long fd_p, const char * name
 }
 
 void adios_get_write_buffer_ (long long * fd_p, const char * name
-                             ,unsigned long long * size
+                             ,uint64_t * size
                              ,void ** buffer, int err, int name_size
                              )
 {
